requirements.txt
Flask==3.1.1
Flask-Session==0.8.0
supabase==2.16.0
python-dotenv==1.1.1
gunicorn==22.0.0